chore(deser-lib): move current code to cbor namespace
TICKET: HSM-236
